Выносит серверный рендеринг в middleware
